+++ src/main/java/org/cookieandkakao/babting/common/config/WebConfig.java
@@ -1,7 +1,7 @@
 package org.cookieandkakao.babting.common.config;
 
 import java.util.List;
-import org.cookieandkakao.babting.domain.member.LoginMemberIdArgumentResolver;
+import org.cookieandkakao.babting.common.resolver.LoginMemberIdArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
